+++ app_wasm.js
@@ -1,5 +1,8 @@
 
-window.onfocus = function(event) {
+window.onfocus = function() {
 	windowFocus = true
 	sendMessage( "session-resume{session=" + sessionID +"}" );
+}
+
+function closeSocket() {
 }